requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
dominate==2.5.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
imageio==2.9.0
importlib-metadata==2.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.1
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.3
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tifffile==2020.9.3
torch==1.4.0
torchvision==0.5.0
tqdm==4.50.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1
